#153, continued ds rework
