+++ cate/ds/local.py
@@ -19,7 +19,9 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-__author__ = "Marco Zühlke (Brockmann Consult GmbH)"
+__author__ = "Norman Fomferra (Brockmann Consult GmbH), " \
+             "Marco Zühlke (Brockmann Consult GmbH), " \
+             "Chris Bernat (Telespazio VEGA UK Ltd)"
 
 """
 Description
@@ -43,7 +45,6 @@ import json
 import os
 import shutil
 import xarray as xr
-from abc import ABCMeta
 from collections import OrderedDict
 from datetime import datetime
 from dateutil import parser
@@ -51,12 +52,14 @@ from glob import glob
 from typing import Optional, Sequence, Union, Any
 from xarray.backends.netCDF4_ import NetCDF4DataStore
 
+from cate.conf import get_config_value
+from cate.conf.defaults import NETCDF_COMPRESSION_LEVEL
 from cate.core.ds import DATA_STORE_REGISTRY, DataStore, DataSource, open_xarray_dataset
 from cate.core.ds import get_data_stores_path
 from cate.core.types import GeometryLike, TimeRange, TimeRangeLike, VariableNamesLike
 from cate.util.misc import to_list
 from cate.util.monitor import Monitor
-
+from cate.ds.config import LocalDataSourceConfiguration
 _REFERENCE_DATA_SOURCE_TYPE = "FILE_PATTERN"
 
 
@@ -68,86 +71,6 @@ def get_data_store_path():
 def add_to_data_store_registry():
     data_store = LocalDataStore('local', get_data_store_path())
     DATA_STORE_REGISTRY.add_data_store(data_store)
-
-
-class LocalDataSourceConfiguration(metaclass=ABCMeta):
-
-    def __init__(self, name: str, data_store: Union[DataStore, str], config_type: str, source: str = None):
-        self._name = "{}.{}".format(data_store, name)
-        self._source = source
-        self._config_type = config_type
-        self._files = OrderedDict()
-        self._data_store = data_store if isinstance(data_store, DataStore) else \
-            DATA_STORE_REGISTRY.get_data_store(data_store)
-
-    @property
-    def name(self):
-        return self._name
-
-    def add_file(self, path: str, time_range: TimeRangeLike.TYPE = None, update: bool = False):
-
-        is_disjoint = self._files.keys().isdisjoint([path])
-        if not update and not is_disjoint:
-            raise ValueError("Config already contains file `{}`".format(path))
-
-        self._files[path] = time_range
-        self._files = OrderedDict(sorted(self._files.items(), key=lambda f: f[1] if f[1] is not None else datetime.max))
-
-    def remove_file(self, path: str):
-        is_disjoint = self._files.keys().isdisjoint([path])
-        if is_disjoint:
-            raise ValueError("Config does not contains file `{}`".format(path))
-        self._files.pop(path)
-
-    def save(self):
-        config = OrderedDict({
-            'name': self._name,
-            'meta_data': {
-                'type': self._config_type,
-                'data_store': self._data_store.name,
-                'temporal_coverage': None,
-                'spatial_coverage': None,
-                'variables': [],
-                'source': self._source,
-                'last_update': None,
-                'last_source_update': None
-            },
-            'files': [[item[0], item[1][0], item[1][1]] for item in self._files.items()]
-        })
-        dump_kwargs = dict(indent='  ', default=self._json_default_serializer)
-        file_name = os.path.join(self._data_store.data_store_path, self._name + '.json')
-        with open(file_name, 'w') as fp:
-            json.dump(config, fp, **dump_kwargs)
-
-    @staticmethod
-    def load(json_path: str) -> Optional['LocalDataSourceConfiguration']:
-        config = None
-        if os.path.isfile(json_path):
-            with open(json_path) as fp:
-                config = json.load(fp=fp) or None
-        if config:
-            datasource_name = config.get('name')
-            meta_data = config.get('meta_data', {})
-            datastore_name = meta_data.get('data_store')
-            temporal_coverage = meta_data.get('temporal_coverage',None)
-            spatial_coverage = meta_data.get('spatial_coverage',None)
-            variables = meta_data.get('variables', [])
-            source = meta_data.get('source', None)
-            last_update = meta_data.get('last_update', None)
-            last_source_update = meta_data.get('last_source_update', None)
-
-            return LocalDataSourceConfiguration(datasource_name, datastore_name, source)
-        return None
-
-    @staticmethod
-    def _json_default_serializer(obj):
-        if isinstance(obj, datetime):
-            return obj.isoformat()
-        raise TypeError('Not sure how to serialize %s' % (obj,))
-
-    @staticmethod
-    def check_type(config_json: dict, expected_type: str):
-        return expected_type.lower() == config_json.get('type', 'unknown').lower()
 
 
 class LocalDataSource(DataSource):
@@ -201,21 +124,29 @@ class LocalDataSource(DataSource):
             return None
 
     def _make_local(self,
-                   local_name: str,
-                   local_id: str = None,
-                   time_range: TimeRangeLike.TYPE = None,
-                   region: GeometryLike.TYPE = None,
-                   var_names: VariableNamesLike.TYPE = None,
-                   monitor: Monitor = Monitor.NONE) -> str:
-
+                    local_name: str,
+                    local_id: str = None,
+                    time_range: TimeRangeLike.TYPE = None,
+                    region: GeometryLike.TYPE = None,
+                    var_names: VariableNamesLike.TYPE = None,
+                    monitor: Monitor = Monitor.NONE) -> str:
+        if not local_id:
+            local_id = local_name
         time_range = TimeRangeLike.convert(time_range) if time_range else None
         region = GeometryLike.convert(region) if region else None
         var_names = VariableNamesLike.convert(var_names) if var_names else None  # type: Sequence
 
-        local_path = os.path.join(get_data_store_path(), local_name)
+        local_path = os.path.join(get_data_store_path(), local_id)
 
-        config = LocalDataSourceConfiguration(name=local_name, data_store=self._data_store,
+        config = LocalDataSourceConfiguration(name=local_id, data_store=self._data_store,
                                               config_type=_REFERENCE_DATA_SOURCE_TYPE, source=self._name)
+
+        compression_level = get_config_value('NETCDF_COMPRESSION_LEVEL', NETCDF_COMPRESSION_LEVEL)
+        compression_enabled = True if compression_level > 0 else False
+
+        encoding_update = dict()
+        if compression_enabled:
+            encoding_update.update({'zlib': True, 'complevel': compression_level})
 
         remote_paths = [glob(file[0] for file in self._files.items())]
         if not remote_paths:
@@ -258,8 +189,6 @@ class LocalDataSource(DataSource):
                                 [var_name for var_name in var_names if var_name != sel_var_name])
                             local_netcdf.store_dataset(var_dataset)
                             child_monitor.progress(work=1, msg=sel_var_name)
-                    except:
-                        raise
                     finally:
                         if remote_netcdf:
                             remote_netcdf.close()
@@ -272,7 +201,7 @@ class LocalDataSource(DataSource):
         config.save()
         monitor.done()
 
-        return local_name
+        return local_id
 
     def make_local(self,
                    local_name: str,
